Migrate Ed25519VRF rhoToHash functionality to the consensus package object
