+++ consensus/src/main/scala/co/topl/consensus/interpreters/ChainSelection.scala
@@ -7,10 +7,10 @@ import cats.implicits._
 import co.topl.algebras.UnsafeResource
 import co.topl.consensus.algebras.ChainSelectionAlgebra
 import co.topl.crypto.hash.Blake2b512
-import co.topl.crypto.signing.Ed25519VRF
 import co.topl.models._
 import co.topl.models.utility._
 import co.topl.consensus.models.SlotData
+import co.topl.consensus.rhoToRhoTestHash
 import co.topl.typeclasses.implicits._
 import org.typelevel.log4cats.Logger
 import org.typelevel.log4cats.slf4j.Slf4jLogger
@@ -33,7 +33,7 @@ object ChainSelection {
     val slotOrder = Order.by[NonEmptyChain[SlotData], Slot](-_.last.slotId.slot)
     val rhoTestHashOrder =
       Order.reverse(
-        Order.by[NonEmptyChain[SlotData], BigInt](h => BigInt(Ed25519VRF.rhoToRhoTestHash(h.last.rho).toArray))
+        Order.by[NonEmptyChain[SlotData], BigInt](h => BigInt(rhoToRhoTestHash(h.last.rho).toArray))
       )
 
     lengthOrder
