+++ consensus/src/test/scala/co/topl/consensus/interpreters/ChainSelectionSpec.scala
@@ -5,13 +5,13 @@ import cats.effect.unsafe.implicits.global
 import cats.implicits._
 import co.topl.algebras.UnsafeResource
 import co.topl.crypto.hash.Blake2b512
-import co.topl.crypto.signing.Ed25519VRF
 import co.topl.models.ModelGenerators._
 import co.topl.models._
 import co.topl.models.utility.HasLength.instances._
 import co.topl.models.utility.Lengths._
 import co.topl.models.utility.{Lengths, Sized}
 import co.topl.consensus.models.{BlockId, SlotData, SlotId}
+import co.topl.consensus.rhoToRhoTestHash
 import co.topl.models.generators.common.ModelGenerators.genSizedStrictByteString
 import com.google.protobuf.ByteString
 import org.scalamock.scalatest.MockFactory
@@ -134,7 +134,7 @@ class ChainSelectionSpec
     val List(rhoX, rhoY) =
       List
         .tabulate(2)(i => Rho(Sized.strictUnsafe[Bytes, Lengths.`64`.type](Bytes(Array.fill[Byte](64)(i.toByte)))))
-        .sortBy(r => BigInt(Ed25519VRF.rhoToRhoTestHash(r.sizedBytes.data).toArray))
+        .sortBy(r => BigInt(rhoToRhoTestHash(r.sizedBytes.data).toArray))
 
     val xSegment = {
       val base = LazyList
